Re-organize pkg/common/util/reconciler.go
